requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
build==1.2.2.post1
    # via pip-tools
click==8.1.8
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
packaging==24.2
    # via build
pip-tools==7.4.1
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyyaml==6.0.2
    # via -r requirements.in
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
